move router agent to control_flow category
