+++ src/nat/agent/register.py
@@ -20,5 +20,4 @@ from .prompt_optimizer import register as prompt_optimizer
 from .react_agent import register as react_agent
 from .reasoning_agent import reasoning_agent
 from .rewoo_agent import register as rewoo_agent
-from .router_agent import register as router_agent
 from .tool_calling_agent import register as tool_calling_agent
